requirements.txt
google-cloud-aiplatform>=1.60.0
google-cloud-storage>=2.14.0
google-cloud-documentai>=2.20.0
